# generated/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  license:
    name: KSTU
  title: Tourist agency
  version: 1.0.0
servers:
- url: http://127.0.0.1/v1
- url: https://fdobrotv.com/v1
paths:
...
  /flights:
    get:
      operationId: listFlights
      parameters:
      - description: How many items to return at one time (max 100)
        explode: true
        in: query
        name: limit
        required: false
        schema:
          format: int32
          maximum: 100
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flights'
          description: A paged array of flights
          headers:
            x-next:
              description: A link to the next page of responses
              explode: false
              schema:
                type: string
              style: simple
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List all flights
      tags:
      - flights
      x-accepts: application/json
      x-tags:
      - tag: flights
    post:
      operationId: createFlight
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlightIn'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flight'
          description: Ok
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Create a flight
      tags:
      - flights
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: flights
  /flights/{id}:
    delete:
      operationId: deleteFlightById
      parameters:
      - description: The id of the resource to delete
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Resource deleted successfully
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Delete specific flight by id
      tags:
      - flights
      x-accepts: application/json
      x-tags:
      - tag: flights
    get:
      operationId: showFlightById
      parameters:
      - description: The id of the resource to retrieve
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flight'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Info for a specific flight
      tags:
      - flights
      x-accepts: application/json
      x-tags:
      - tag: flights
  /carModels:
    get:
      operationId: listCarModels
      parameters:
      - description: How many items to return at one time (max 100)
        explode: true
        in: query
        name: limit
        required: false
        schema:
          format: int32
          maximum: 100
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarModels'
          description: A paged array of car models
          headers:
            x-next:
              description: A link to the next page of responses
              explode: false
              schema:
                type: string
              style: simple
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List all car models
      tags:
      - carModels
      x-accepts: application/json
      x-tags:
      - tag: carModels
    post:
      operationId: createCarModel
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarModelIn'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarModel'
          description: Ok
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Create a car model
      tags:
      - carModels
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: carModels
  /carModels/{id}:
    delete:
      operationId: deleteCarModelById
      parameters:
      - description: The id of the resource to delete
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Resource deleted successfully
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Delete specific car model by id
      tags:
      - carModels
      x-accepts: application/json
      x-tags:
      - tag: carModels
    get:
      operationId: showCarModelById
      parameters:
      - description: The id of the resource to retrieve
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarModel'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Info for a specific car model
      tags:
      - carModels
      x-accepts: application/json
      x-tags:
      - tag: carModels
  /carMarks:
    get:
      operationId: listCarMarks
      parameters:
      - description: How many items to return at one time (max 100)
        explode: true
        in: query
        name: limit
        required: false
        schema:
          format: int32
          maximum: 100
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarMarks'
          description: A paged array of car marks
          headers:
            x-next:
              description: A link to the next page of responses
              explode: false
              schema:
                type: string
              style: simple
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List all carMarks
      tags:
      - carMarks
      x-accepts: application/json
      x-tags:
      - tag: carMarks
    post:
      operationId: createCarMark
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarMarkIn'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarMark'
          description: Ok
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Create a car mark
      tags:
      - carMarks
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: carMarks
  /carMarks/{id}:
    delete:
      operationId: deleteCarMarkById
      parameters:
      - description: The id of the resource to delete
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Resource deleted successfully
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Delete specific car by id
      tags:
      - carMarks
      x-accepts: application/json
      x-tags:
      - tag: carMarks
    get:
      operationId: showCarMarkById
      parameters:
      - description: The id of the resource to retrieve
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarMark'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Info for a specific car mark
      tags:
      - carMarks
      x-accepts: application/json
      x-tags:
      - tag: carMarks
  /cars:
    get:
      operationId: listCars
      parameters:
      - description: How many items to return at one time (max 100)
        explode: true
        in: query
        name: limit
        required: false
        schema:
          format: int32
          maximum: 100
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cars'
          description: A paged array of cars
          headers:
            x-next:
              description: A link to the next page of responses
              explode: false
              schema:
                type: string
              style: simple
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List all cars
      tags:
      - cars
      x-accepts: application/json
      x-tags:
      - tag: cars
    post:
      operationId: createCar
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarIn'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
          description: Ok
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Create a car
      tags:
      - cars
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: cars
  /cars/{id}:
    delete:
      operationId: deleteCarById
      parameters:
      - description: The id of the resource to delete
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Resource deleted successfully
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Delete specific car by id
      tags:
      - cars
      x-accepts: application/json
      x-tags:
      - tag: cars
    get:
      operationId: showCarById
      parameters:
      - description: The id of the resource to retrieve
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Info for a specific car
      tags:
      - cars
      x-accepts: application/json
      x-tags:
      - tag: cars
...
components:
  schemas:
...
    CarColor:
      enum:
      - white
      - black
      - brown
      - yellow
      - red
      - blue
      - silver
      type: string
...
    Flights:
      items:
        $ref: '#/components/schemas/Flight'
      maxItems: 100
      type: array
    FlightIn:
      example:
        departureAirport: Moscow
        departureDateTime: 2000-01-23T04:56:07.000+00:00
        arrivalDateTime: 2000-01-23T04:56:07.000+00:00
        arrivalAirport: Antalia
      properties:
        departureAirport:
          example: Moscow
          type: string
        arrivalAirport:
          example: Antalia
          type: string
        departureDateTime:
          format: date-time
          type: string
        arrivalDateTime:
          format: date-time
          type: string
      required:
      - arrivalAirport
      - arrivalDateTime
      - departureAirport
      - departureDateTime
      type: object
    Flight:
      allOf:
      - $ref: '#/components/schemas/FlightIn'
      - properties:
          id:
            format: uuid
            type: string
        required:
        - id
        type: object
...
    CarMarks:
      items:
        $ref: '#/components/schemas/CarMark'
      maxItems: 100
      type: array
    CarMarkIn:
      example:
        name: Peugeot
      properties:
        name:
          example: Peugeot
          type: string
      required:
      - name
      type: object
    CarMark:
      allOf:
      - $ref: '#/components/schemas/CarMarkIn'
      - properties:
          id:
            format: uuid
            type: string
        required:
        - id
        type: object
    CarModels:
      items:
        $ref: '#/components/schemas/CarModel'
      maxItems: 100
      type: array
    CarModelIn:
      example:
        name: Traveller
      properties:
        name:
          example: Traveller
          type: string
      required:
      - name
      type: object
    CarModel:
      allOf:
      - $ref: '#/components/schemas/CarModelIn'
      - properties:
          id:
            format: uuid
            type: string
        required:
        - id
        type: object
    Cars:
      items:
        $ref: '#/components/schemas/Car'
      maxItems: 100
      type: array
    CarIn:
      example:
        markId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        color: null
        modelId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        plateNumber: A123TH716RUS
      properties:
        markId:
          format: uuid
          type: string
        modelId:
          format: uuid
          type: string
        plateNumber:
          example: A123TH716RUS
          type: string
        color:
          $ref: '#/components/schemas/CarColor'
      required:
      - color
      - markId
      - modelId
      - plateNumber
      type: object
    Car:
      example:
        color: null
        model: Traveller
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        plateNumber: A123TH716RUS
        mark: Peugeot
      properties:
        id:
          format: uuid
          type: string
        mark:
          example: Peugeot
          type: string
        model:
          example: Traveller
          type: string
        plateNumber:
          example: A123TH716RUS
          type: string
        color:
          $ref: '#/components/schemas/CarColor'
      required:
      - color
      - id
      - mark
      - model
      - plateNumber
      type: object
...
    Error:
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
      required:
      - code
      - message
      type: object